% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\alias{ps_has_key}
\alias{ps_get_key}
\title{Check or Get PlanScore API Key}
\usage{
ps_has_key()

ps_get_key()
}
\value{
logical if \code{has}, key invisibly if \code{get}
}
\description{
Check or Get PlanScore API Key
}
\examples{
ps_has_key()
}
\concept{key}
