% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{ps_capture}
\alias{ps_capture}
\title{Capture PlanScore Graphs}
\usage{
ps_capture(link, path)
}
\arguments{
\item{link}{plan_url output from \code{ps_upload()} functions}

\item{path}{path to save copy of graphs in, likely ending in \code{.png}}
}
\value{
path to screenshot
}
\description{
Capture PlanScore Graphs
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# often times out
url <- 'https://planscore.org/plan.html?20221127T213653.168557156Z'
tf <- tempfile(fileext = '.png')
ps_capture(url, path = tf)
\dontshow{\}) # examplesIf}
}
\concept{results}
