% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ps_ingest}
\alias{ps_ingest}
\title{Ingest PlanScore Output}
\usage{
ps_ingest(link)
}
\arguments{
\item{link}{index url output from \code{ps_upload} functions}
}
\value{
tibble with district and plan level data
}
\description{
Ingest PlanScore Output
}
