% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{ps_set_key}
\alias{ps_set_key}
\title{Add Entry to Renviron}
\usage{
ps_set_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Character. API key to add to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{PLANSCORE_KEY} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
key, invisibly
}
\description{
Adds PlanScore API key to .Renviron.
}
\examples{
\dontrun{
set_planscore_key('1234')
}

}
\concept{key}
