% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_states.R
\name{ps_supported_states}
\alias{ps_supported_states}
\title{List Currently Supported States}
\usage{
ps_supported_states()
}
\value{
tibble of supported states
}
\description{
List Currently Supported States
}
\examples{
ps_supported_states()
}
\concept{info}
