% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ps_ingest}
\alias{ps_ingest}
\title{Ingest PlanScore Output}
\usage{
ps_ingest(link)
}
\arguments{
\item{link}{index url output from \code{ps_upload()} functions}
}
\value{
\code{tibble} with district and plan level data
}
\description{
Ingest PlanScore Output
}
\examples{
url <- 'https://planscore.s3.amazonaws.com/uploads/20221127T213653.168557156Z/index.json'
ps_ingest(url)
}
\concept{results}
