% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ps_save}
\alias{ps_save}
\title{Save PlanScore Output}
\usage{
ps_save(link, path, json = TRUE)
}
\arguments{
\item{link}{index url output from \code{ps_upload()} functions}

\item{path}{path to save copy of data in, likely ending in \code{.json} if \code{json = TRUE}
or \code{.tsv} if \code{json = FALSE}.}

\item{json}{should the file be saved as \code{.json} (\code{TRUE}) or \code{.tsv} (\code{FALSE}).
\code{json = TRUE} is slower but contains more information.}
}
\value{
path to \code{json} file
}
\description{
Save PlanScore Output
}
\examples{
url <- 'https://planscore.s3.amazonaws.com/uploads/20221127T213653.168557156Z/index.json'
tf <- tempfile(fileext = '.json')
ps_save(url, tf)
}
\concept{results}
