% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{ps_upload_file}
\alias{ps_upload_file}
\title{Title}
\usage{
ps_upload_file(file, description, temporary = TRUE)
}
\arguments{
\item{file}{file to upload}

\item{temporary}{Should a temporary PlanScore upload be used? Default is TRUE.}
}
\description{
Title
}
\examples{
\dontshow{if (PlanScoreR::ps_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API Key

\dontshow{\}) # examplesIf}
}
